Modify visualization of tracks
Modify track projections and both 2D and 3D event visualization so that the track points are proportional to the deposited energy instead of the shower size. This way, the point size illustrates the light production. This is important for ascending showers, where the shower size keeps approximatey constant in high atmosphere, although the light production decreases rapidly because of the low density.