Fix issues with projection plots
Include the first interaction point and the exit point of the shower (either at ground or at the top of the atmosphere) in projection plots so that it can be seen the entire track even tough the number of track points is small.
Use nadir projection when the telescope is (mostly) looking down.